Add files via upload
R code for predictive model :

Using SVM with various kernels and with Weight of evidence for categorical variables. ( Got  Average Accuracy of ~60% )

Using Randomforest with 500 and 1000 trees and with Weight of evidence for categorical variables. ( Got Accuracy of ~80% )

Using logistic regression with Weight of evidence for categorical variables. ( But it gives warning - Need to look at it )